Use new file for resources_icon + dark theme improvments
